it works!

+++ Assign2Minesweeper/Assign2Minesweeper/Typedefs.h
@@ -4,6 +4,6 @@
 #include <vector>
 typedef signed char byte;
 typedef unsigned char ubyte;
-typedef std::pair<int,int> point;
+typedef std::pair<size_t,size_t> point;
 typedef std::vector<std::vector<byte>> mineFieldData;
 #endif
